Add way to retrieve rx_queue mempool
This allows allocating packets from the same mempool. Since only a
single mempool is allowed for use with each tx_queue we need to retrieve
some packets from that pool to send original packets. When the first
packet arrives we could inspect it for its Mempool anyways, however,
this is not practical at all.